% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_model.R
\name{cumulative_model}
\alias{cumulative_model}
\title{Fit a model to a species using the cumulative predictions for a secundary
species}
\usage{
cumulative_model(
  species = "Adal_dece",
  min_occurrences = 1000,
  min_species = 3,
  secondary,
  first_order = TRUE
)
}
\arguments{
\item{species}{Name of the species.}

\item{min_occurrences}{The minimum number of occurrences per species.}

\item{min_species}{The minimum number of species recorded at the combination
of location and year.}

\item{secondary}{The output of \code{base_model()} for a different species.}

\item{first_order}{Use first (\code{TRUE}) or second (\code{FALSE}) order random walk
for the year component.
Defaults to \code{TRUE}.}
}
\description{
Fit a model to a species using the cumulative predictions for a secundary
species
}
