% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a model to a species using the predictions for a secondary species}
\usage{
fit_model(
  base_data,
  trend_prediction,
  base_prediction,
  field_prediction,
  knots
)
}
\arguments{
\item{base_data}{A dataframe with the base data.}

\item{trend_prediction}{A dataframe with the timestamps to predict the trend.}

\item{base_prediction}{A dataframe with the locations and timestamps to
predict.}

\item{field_prediction}{A dataframe with locations and timestamps to predict
the spatiotemporal field.}

\item{knots}{Which years to use a knots for the piecewise linear regression.}
}
\description{
Fit a model to a species using the predictions for a secondary species
}
