% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_model.R
\name{get_country_grid}
\alias{get_country_grid}
\title{Get a grid of points within the country}
\usage{
get_country_grid(
  path = ".",
  country = c("BE", "GB", "NL", "all"),
  cellsize = 10000,
  what = "polygons"
)
}
\arguments{
\item{path}{Location of the Natural Earth Data.
The function will download and unzip the data when not available.}

\item{country}{Data from which country to select.}

\item{cellsize}{Size of the grid cells in meter.
Must be at least 1000 m.
Defaults to 10.000 m.}

\item{what}{character; one of: \code{"polygons"}, \code{"corners"}, or \code{"centers"}}
}
\description{
We use the borders from \href{https://www.naturalearthdata.com/}{Natural Earth Data}
}
