% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence_map.R
\name{occurrence_map}
\alias{occurrence_map}
\title{Display a leaflet map with the occurrences for a given species}
\usage{
occurrence_map()
}
\description{
Display a leaflet map with the occurrences for a given species
}
