% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_smoother_results.R
\name{prepare_smoother_results}
\alias{prepare_smoother_results}
\title{Prepare the results from the smoother models for the report}
\usage{
prepare_smoother_results(models = ".", maps = ".", output = ".")
}
\arguments{
\item{models}{the path to the individual model results}

\item{maps}{Location of the Natural Earth Data.
See \code{get_country_grid()}.}

\item{output}{the path for the output file}
}
\description{
Prepare the results from the smoother models for the report
}
