% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_model.R
\name{probability_model}
\alias{probability_model}
\title{Fit a model to a species using the predictions for a secondary species}
\usage{
probability_model(
  species = "Adal_bipu",
  min_occurrences = 1000,
  min_species = 3,
  secondary,
  knots = c(1990, 2000, 2010, 2020),
  country = c("BE", "NL")
)
}
\arguments{
\item{species}{Name of the species.}

\item{min_occurrences}{The minimum number of occurrences per species.}

\item{min_species}{The minimum number of species recorded at the combination
of location and year.}

\item{secondary}{The output of \code{base_model()} for a different species.}

\item{knots}{Which years to use a knots for the piecewise linear regression.}

\item{country}{Data from which country to select.}
}
\description{
Fit a model to a species using the predictions for a secondary species
}
