% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_model.R
\name{secondary_model}
\alias{secondary_model}
\title{Fit a model to a species using the predictions for a secundary species}
\usage{
secondary_model(
  species = "Adal_dece",
  min_occurrences = 1000,
  min_species = 3,
  secondary
)
}
\arguments{
\item{species}{Name of the species}

\item{min_occurrences}{The minimum number of occurrences per species.}

\item{min_species}{The minimum number of species recorded at the combination
of location and year.}

\item{secondary}{the output of \code{base_model()} for a different species}
}
\description{
Fit a model to a species using the predictions for a secundary species
}
