% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_model.R
\name{smooth_model}
\alias{smooth_model}
\title{Fit a base model to a species}
\usage{
smooth_model(
  species = "Harm_axyr",
  min_occurrences = 1000,
  min_species = 3,
  country = c("BE", "GB", "NL", "all"),
  path = ".",
  cellsize = 10000,
  buffer_distance = 50000,
  buffer_locations = 50
)
}
\arguments{
\item{species}{Name of the species.}

\item{min_occurrences}{The minimum number of occurrences per species.}

\item{min_species}{The minimum number of species recorded at the combination
of location and year.}

\item{country}{Data from which country to select.}

\item{path}{Location of the Natural Earth Data.
The function will download and unzip the data when not available.}

\item{cellsize}{Size of the grid cells in meter.
Must be at least 1000 m.
Defaults to 10.000 m.}

\item{buffer_distance}{The range in which to look for other locations.}

\item{buffer_locations}{The minimal number of locations with suffient data
within the buffer range}
}
\description{
Fit a base model to a species
}
