% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{base_model}
\alias{base_model}
\title{Fit a base model to a species}
\usage{
base_model(
  species = "Harm_axyr",
  min_occurrences = 1000,
  min_species = 3,
  knots = c(1990, 2000, 2010, 2020),
  country = c("BE", "NL")
)
}
\arguments{
\item{species}{Name of the species.}

\item{min_occurrences}{The minimum number of occurrences per species.}

\item{min_species}{The minimum number of species recorded at the combination
of location and year.}

\item{knots}{Which years to use a knots for the piecewise linear regression.}

\item{country}{Data from which country to select.}
}
\description{
Fit a base model to a species
}
