% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{base_model}
\alias{base_model}
\title{Fit a base model to a species}
\usage{
base_model(
  species = "Harm_axyr",
  min_occurrences = 1000,
  min_species = 3,
  first_order = TRUE,
  center_year = 2001
)
}
\arguments{
\item{species}{Name of the species.}

\item{min_occurrences}{The minimum number of occurrences per species.}

\item{min_species}{The minimum number of species recorded at the combination
of location and year.}

\item{first_order}{Use first (\code{TRUE}) or second (\code{FALSE}) order random walk
for the year component.
Defaults to \code{TRUE}.}

\item{center_year}{The year to center to.
Defaults to \code{2001}.}
}
\description{
Fit a base model to a species
}
