% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a model to a species using the predictions for a secondary species}
\usage{
fit_model(first_order = TRUE, base_data, trend_prediction, base_prediction)
}
\arguments{
\item{first_order}{Use first (\code{TRUE}) or second (\code{FALSE}) order random walk
for the year component.
Defaults to \code{TRUE}.}

\item{base_data}{A dataframe with the base data.}

\item{trend_prediction}{A dataframe with the timestamps to predict the trend.}

\item{base_prediction}{A dataframe with the locations and timestamps to
predict.}
}
\description{
Fit a model to a species using the predictions for a secondary species
}
