% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import and standardise the raw data}
\usage{
import_data(
  belgium_occurrence,
  belgium_visits,
  netherlands_occurrence,
  netherlands_visits,
  gb_occurrence,
  gb_visits,
  output,
  strict = TRUE,
  latest_year = 2020
)
}
\arguments{
\item{belgium_occurrence}{path to the CSV file with the Belgian occurrence
data.}

\item{belgium_visits}{path to the CSV file with the Belgian visit data.}

\item{netherlands_occurrence}{path to the CSV file with the Dutch occurrence
data.}

\item{netherlands_visits}{path to the CSV file with the Dutch visit data.}

\item{gb_occurrence}{path to the CSV file with the British occurrence
data.}

\item{gb_visits}{path to the CSV file with the British visit data.}

\item{output}{path to the root of the data package}

\item{strict}{What to do when the metadata changes. \code{strict = FALSE}
overwrites the data and the metadata with a warning listing the changes,
\code{strict = TRUE} returns an error and leaves the data and metadata as is.
Defaults to \code{TRUE}.}

\item{latest_year}{Don't import data more recent than this year.
Defaults to 2020.}
}
\description{
Import and standardise the raw data
}
