% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import and standardise the raw data}
\usage{
import_data(belgium, output)
}
\arguments{
\item{belgium}{path to the CSV file with the Belgian data.}

\item{output}{path to the root of the data package}
}
\description{
Import and standardise the raw data
}
