% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_relevant.R
\name{load_relevant}
\alias{load_relevant}
\title{Load the relevant occurrence data}
\usage{
load_relevant(
  min_occurrences = 1000,
  min_species = 3,
  country = c("BE", "GB", "NL", "all"),
  buffer_distance = 50000,
  buffer_locations = 50
)
}
\arguments{
\item{min_occurrences}{The minimum number of occurrences per species.}

\item{min_species}{The minimum number of species recorded at the combination
of location and year.}

\item{country}{Data from which country to select.}

\item{buffer_distance}{The range in which to look for other locations.}

\item{buffer_locations}{The minimal number of locations with suffient data
within the buffer range}
}
\description{
Load the relevant occurrence data
}
