% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_smoother_results.R
\name{prepare_smoother_results}
\alias{prepare_smoother_results}
\title{Prepare the results from the smoother models for the report}
\usage{
prepare_smoother_results(models = ".", maps = ".", output = ".")
}
\arguments{
\item{models_path}{the path to the individual model results}
}
\description{
Prepare the results from the smoother models for the report
}
